logging.conf: exchanged for internal configuration
Python is well known for it's unintuitive treatment of filesystem calls
relative to execution instead of written occurence. With
our `logging.conf` we had such an issue. We believed putting a
`logging.conf` in the same folder as devicetestcase.py and adding a call to
this `logging.conf` would be enough. But of course, when importing the
devicetestcase somewhere else, it will try to find the the `logging.conf`
in the other folder. After considering some different options, we decided
to put the configuration directly into the devicetestcase.py. It is not
all too big and we felt this way there is the smallest chance to make
errors.

Original commit by C.Bruhn <bruhn@dresearch-fe.de> and
M.Schuknecht <schuknecht@dresearch-fe.de>. Restructured by E.Bernoth.

This commit belongs to the Issue #9.
Also see 873a4a8e427d40109ef146f90962b8502a198430

Signed-off-by: Erik Bernoth <bernoth@dresearch-fe.de>
Acked-by: Mario Schuknecht <schuknecht@dresearch-fe.de>
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>