moved src files to new src folder
It is not common for a professional python project to put all it's source files
in it's root folder. Thus a subfolder `src` was created.
Another subfolder, called `monk_tf`, was created underneath that Python can
recognize our code as a module. This makes sense if we ever decide that this
project needs to be seperated into more submodules, thus it's not a bad idea.
It will also help with adding distutils/pypi support for this project.

Signed-off-by: Erik Bernoth <bernoth@dresearch-fe.de>