environment variable for no reboot at start
The devicetestcase.py is changed in order to not do a reboot if it is not
necessary. To configure the reboot at the start of a devicetestcase you
use the environment variable "MONK_NEED_START_REBOOT". Value 1 means
reboot necessary and 0 means reboot not necessary. In default the device
is rebooting at start of the test.

Signed-off-by: Johannes Kroop <johkroop@gmail.com>
Acked-by: Erik Bernoth <bernoth@dresearch-fe.de>
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>