[doc] new page contributing
which describes the development process and ways to interact with the
MONK community.

See Github Issue #28 for more info.
Also some of the changes are more about the policies, thus look at Github Issue
 #30 for more details.

This change also includes the changes from Github Issue #21, which added some
info about the environment variables configuration to the old docs.

Signed-off-by: Erik Bernoth <bernoth@dresearch-fe.de>
Acked-by: Angelika Tobisch <tobisch@dresearch-fe.de>
Acked-by: Andrea Schönfeldt <schoenfeldt@dresearch-fe.de>