docs: changed pictures
Also changed how the pictures are used in the index.rst file. Now they are
centered and minimized a little. This way it is more clear to see what is the
label for the figures, which again makes it possible to remove the "figure x:"
text before the figure labes. Removing them makes the PDF labeling more
beautiful.

See Github Issue #40 for more details.

Pictures created by Zhang, Mengqiao. Update to color space by Binschek, Eik.

Signed-off-by: Mengqiao Zhang <mengqiao.zhang@gmail.com>
Signed-off-by: Eik Binschek <binschek@dresearch-fe.de>
Acked-by: Erik Bernoth <bernoth@dresearch-fe.de>
Acked-by: Johannes Kroop <kroop@dresearch-fe.de>