Codingstyle: New Guide
Adding the first public style guide for MONK, as derived from the Team Leonardo
of DResearch Fahrzeugeletronik GmbH. Because it was rather hard to identify the
meaning of some of the rules defined in the TeamL guide the decision was made
to formulate the rules in a more expressive manner, with reasons, sources and
examples when possible. This might have also resulted in minor changes to the
rules here and there. It seemed meaningful in some places to repeat already
said rules or example code snippets, because it is likely that the guide will
be used as a reference more often then read as a whole and having important
things together at the same place seemed more helpful in that regard than
avoiding repetition.

See Issue #4.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Angelika Tobisch <tobisch@dresearch-fe.com>
Acked-by: Steffen Sledz <sledz@dresearch-fe.com>