serial_conn.py: use of parameter timeout
This is a change to fix the bug # 39. The Issue for that bug is incorrectly
described. The Parameter timeout is already in use, but only for reading
access. After that fix it is aso in use for writing access. The class
SerialConn inherits from Serial. The timeouts are regulated with the variables
_timeout for reading access and _writeTimeout for writing access.
See line 75-79 in serial_conn:

    self._writeTimeout = timeout
    self._timeout = timeout

See also:
Issue #39

Signed-off-by: Johannes Kroop <johkroop@gmail.com>
Acked-by: Erik Bernoth <bernoth@dresearch-fe.de>
Acked-by: Eik Binschek <binschek@dresearch-fe.de>
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>