serial_conn.py: Added unit tests for timeouts.
Added 2 tests. One for the timeout, the other for the writeTimeout properties
used in serial_conn.SerialConn.read_until() and a mock for the pySerial API.

Getting the tests to work required changing the serial_conn as well. The
protected variables shouldn't be used, if not explicitely necessary. Instead
using the property is better and public API compliant.

Smaller cleanups:
 - changed the if, because a check for `txt == ""` is the same as a test for
   `not txt`.
 - if one local var is called `old_wto` it makes sense to call the other one
   `old_rto` as well

See Github Issue #39.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Signed-off-by: Johannes Kroop <kroop@dresearch-fe.de>
Acked-by: Eik Binschek <binschek@dresearch-fe.de>
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>