.travis.yml: Update for simpler unittesting
Because the interface for testing and installing did once more got closer to
the standard, there is already a payoff in the travis-ci configuration, were
now less lines of code are needed.

In the new configuration the setuptools standard interface is used for
installation as well as testing. The only requirements necessary for
installation left are python2.7-dev and libssh2-1-dev, which are C-header
dependencies needed for pylibssh2 to build. As far as I see there is not much
we can do about those requirements for now.

See Issue #34.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Eik Binschek <binschek@dresearch-fe.de>
Acked-by: Johannes Kroop <kroop@dresearch-fe.de>