development-env.txt added: Sphinx in virtualenvs
There was a problem using globally installed Sphinx in virtualenvs. There's no
problem though, if it was installed with pip inside of a virtualenv.

At first it was considered to add sphinx to either `install_requires`,
`extras_require`, `setup_requires`, `tests_require` in the setup.py file. But
neither seems to be a good use case for building documentation. A variable like
`development_requires` doesn't exist. Therefore something like
`development_requires` was initiated with a development-env.txt. This contains
the state of the current development environment inside virtualenv. It can be
installed like that:

   # make sure you are inside the virtualenv
   $ source bin/activate
   # install the requirements from the file (note the -r argument)
   $ pip install -r development-env.txt

If you get errors from the pip install command, then you are probably missing
the development packages for python and libssh2. Do one of these:

    # Ubuntu
    $ sudo apt-get install python-dev libssh2-1-dev
    # Fedora
    $ sudo yum install python-devel libssh2-devel
    # Sorry, don't know how to do it on Suse

See Github Issue #46.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Eik Binscheck <binschek@dresearch-fe.de>
Acked-by: Johannes Kroop <kroop@dresearch-fe.de>