contributing.rst: Rewrite for development env
Following the recent changes in the development environment, adding support for
virtualenv, depencies and so on, it was a good time to also rewrite the
contributing page. A tutorial was added, which walks through the creation of a
development environment and the creation of this task.

Reviewers should add themselves to the tutorial section which talks about
Acked-bys, to complete the tutorial (there's a TODO for it at the right place)
and they might also explain something about their review process.

See Github Issue #36.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Eik Binschek <binschek@dresearch-fe.de>
Acked-by: Peter Winkler <winkler@dresearch-fe.de>