requirements-*.txt: some minor changes for Jenkins
To apply the new test environment to our Jenkins job, some minor changes had to
be made. A setuptools command for pylint was installed, that pylint can be
started like all the commands via `python setup.py <command>`. This will
automatically add pylint itself to the requirements.

For testing and for developing there are now 2 requirement files:

 * requirements-develop.txt
 * requirements-test.txt

Advantage for this is, that everything is installed with pip and everybody
should use the same versions for their tests/development env. Disadvantage is
that from time to time someone needs to update those files.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Steffen Sledz <sledz@dresearch-fe.de>
Acked-by: Christina Quast <quast@dresearch-fe.de>