Jenkins config: Don't Delete Workspace
There was the idea to delete the workspace before or after each Jenkins build.
Another idea got broken by that feature, though. That is the idea, to poll the
GitHub repository every 5 minutes and only build, when a new commit was added
to the watched branch. With deleting the workspace, Jenkins always clones a
completely new repository every time he polls. The result is, that every 5
minutes Jenkins starts a new build. This is not what we desired, and the
functionality of building from clean workspaces wasn't deemed so important.
Therefore we decided to neither clean the workspace before nor after building.

See GitHub Issue #59.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>