conn.py: Implementing connection layer
The connection layer is the lowest abstraction layer of the MONK framework.
Its task is to interact directly with different types of connections. Some
logic is implemented via the State design pattern, and a cmd() method is
implmented as general interface for all connections. This cmd() method sends
a shell command to the target device and returns its standard output and some
additional information, e.g. the prompt or the return code of the shell
command.

See GitHub Issue #72

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Eik Binschek <eik.binschek@dresearch-fe.de>
Acked-by: Peter Winkler <peter.winkler@dresearch-fe.de>