dev.py: Implementation of the device layer
The device layer is the central layer of the MONK framework. Its task is to
abstract a target device in an object, offering the same cmd() interface as the
connection layer, while handling more errors and many different connections for
the user. By using this layer it should be possible to focus on the interaction
with the target device without worrying how the object and the device actually
communicate with each other.

See GitHub Issue #72

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Eik Binschek <eik.binschek@dresearch-fe.de>
Acked-by: Peter Winkler <peter.winkler@dresearch-fe.de>