conn.py: SerialConn Improved Parameter Parsing
There was a problem with reading harness files that all the parameters are
strings. It was not possible to represent numbers as such (at least no simple
way without changing the grammar etc) or the tuple required for
conn.SerialConnection's login credentials.

The chosen solution is to accept all parameters as strings and transform them
in the conn.SerialConnection's constructor, as well as accepting "user" and
"password" parameters that will be transformed to the credentials form in the
constructor.

Additionally the Exceptionhandling was improved in cases were an incorrect
logger was used and for the use case of being not connected physically to the
serial ports, and the use case in which the executing user does not have the
privileges required by the system.

See GitHub Issue #79.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Eik Binschek <binschek@dresearch-fe.de>