API docs: In depth explanation of State pattern
ATH advised me that some parts of the API docs might be unclear if the State
design pattern is not understood correctly. Therefore I explained a little more
in the connection layer description what actually happens. This might not be
the final result, though. We might not get around explaining all the design
principles at some point.

See GitHub Issue #83.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>
Acked-by: Angelika Tobisch <tobisch@dresearch-fe.de>