fixture.py: provide a simple cmd() interface
I assume that most test cases will only work with a single device, therefore
having an interface, like:

    fixt.devs[0].cmd(...)

Is awful! It is better to provide this as a single method directly in the
fixture object, like:

    fixt.cmd(...)

which does the exact same thing.

This is what is implemented here.

See GitHub Issue #93.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>