fixture.py: IOError when fixture file not found
In case of spelling errors fixture should inform the user that a given fixture
file name does not correspond with the existing filestructure. Instead of
finding an individual solution the one implemented in ConfigObj was activated.
This means that if a fixture file cannot be found XiniParser raises an IOError.

See GitHub Issue #95.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>