SerialConn: check if port actually exists
If a given port does not exist, MONK did not respond with any kind of error
message. Now it says that the port is not correct.

For this purpose SerialConn got a new method _is_serialport which checks
whether the given port is a usable tty, and uses this method before keeping the
value of the constructor parameter 'port'. If it is not even possible to open
the given file name, the OSError is caught and False is returned. It seems more
reasonable to respond with return False in this case, because the task of the
method is to check the file name. People might actually call this method
specifically to not run into OSErrors.

See GitHub Issue #100.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>