Disconnected: don't connect without I/O
While using MONK it feels strange to see a SerialConnection move to the state
Connected even when the target device is still deactivated. Therefore the state
Disconnected now asks for a new prompt after a connection was "established" and
if nothing is received (=_prompt raises EmptyResponseException) the state stays
at Disconnected and raises a CantConnectException.

See GitHub Issue #100.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>