SerialConn: take emptiness check out of _cmd()
Having a _cmd() method check for an empty response and raise a corresponding
Exception was a bad design decision, because it adds on the burden of a
developer that might create their own connection type.

Therefore this check and corresponding param was taken out of _cmd() and needs
to be handled elsewhere. The only place where this feature was used until now
was the _prompt() method. Therefore this method gots a corresponding check
added.

See GitHub Issue #100

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>