EchoConnection: explicit "empty" response
It happens that the EchoConnection gets a _cmd() request with an empty command
message, e.g., in the case of being called by _prompt(). However it is unlikely
that there might ever be a case where EchoConnection actually should respond
with an empty string, because such a string is often treated the same way as
`None`. Therefore the decision was made to explicitely return "<empty>" in
these cases.

See GitHub Issue #100

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>