AConnection.can_login(): Only Check Login Prompt
Until now the can_login Method checked for both password and login prompt. This
seems unreasonable. Either the content or the method's name had to be changed.
For now a change of content was assumed the better solution, because the method
was used in several places already.

Also for usability the method was transformed into a property.

See GitHub Issue #104.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>