AConnection._prompt(): Exception depends on prompt
At first _prompt() raised an EmptyResponseException always when an empty string
was the result of its _cmd() call. But that is actually bad in the most common
scenario in which _prompt() basically asks for a current prompt string.
Updating the last_prompt string does not result in any output. Therefore
_prompt() raised an exception in the case where everything worked as expected.

The problem was resolved by also checking the last_prompt attribute. It gets
deleted before the _cmd() call and checked for existance after the call.
Therefore it must have been set in the middle. Only if that didn't happen and
no output was generated then an exception is raised.

FYI: The EmptyResponseException is important for the Disconnected state to
verify if a switch to the Connected state is possible. If you ask for a new
prompt and recieve absolutely nothing back than it is likely that the device is
not running or not connected to the test host properly. Therefore catching an
EmptyResponseException results in not switching to the Connected state.

See GitHub Issue #104.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>