SerialConn: be more verbose about auth. process
Authenticating a user requires actually two steps, sending the username and
then sending the password. While debugging it was not clear whether one of
these steps or something in the middle might result in a buggy behaviour. It
was not the case, but to clear up the code helped evaluating the situation
better.

See GitHub Issue #104.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>