AConnection: Also check for existing attribute
In commit a827b146648 _prompt() got the additional check for the last_prompt
attribute. The goal was to make sure an EmptyResponseException is not just
raised when the _cmd() call does not return any output because that
also happens when the prompt is updated as expected. The new if should only
result in the exception path if last_prompt and output where empty.

I completely forgot, though, that I delete last_prompt. So if it is not set it
simply does not exist and the Exception shown to the user will be an
AttributeError.

Here I correct this by checking for the existance of the last_prompt attribute
first. Only then the exception is raised.

The SilentConnection loses the line where it creates an empty last_prompt
attribute to reflect the change of understanding in the test suite.

This is a late additional commit to GitHub Issue #104.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>