AConnection: raise Exception when not able to auth
If a login() method just returns False when a login didn't succeed the rest of
the code does not know that a problem occured and then just continues to try to
interact with the target device. Therefore a CantAuthException was introduced
in this commit.

This is a first step to resolve GitHub Issue #99.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>