AConnection: check first if already authed
One scenario in which an authentication process might fail is when the user is
already authenticated. Therefore before anything else such a test should be
conducted. Here it is implemented in a seperate property which returns True or
False depending on what the shell command "whoami" returns. Only when the
response to "whoami" is the same as the given username in the credentials
attribute then the return value is True, otherwise False.

See GitHub Issue #99.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>