AConnection: login reacts to pw prompt
It is possible that as the result of what happened before the login process
actually starts with a password prompt by the connection. This happens for
example, when a is_authenticated request or a _prompt() request is interpreted
as a username by the target device. This is actually what happens in the most
likely scenario of connecting to the device with MONK.

This scenario is treated by checking for it with the previously called
can_auth() and then request a _prompt() when it is likely that the target
device expects a password not a username. This was probably also the reason why
can_auth() tested for user prompt and pw prompt.

Now that I understand the initial idea behind the code I think renaming the
property can_auth to has_user_prompt is better and adding a second one called
has_pw_prompt for the password prompt. Thereby the if-try-if in the login()
method gets easier to read and is less error prone. Now it says that it only
asks for a new prompt when the old one is the password prompt and then only
attempts a login when the given prompt actually is the user prompt.

See GitHub Issue #99.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>