conn.py: 2 new Prompts on login()
When login via SerialConnection the fist step is to check whether the
connection is already authenticated which employs a "whoami" shell command. If
the conneciton is not authenticated though, the "whoami" is considered the user
name. Because that name usually is no user and therefore has no password, the
password prompt is approached with a new prompt request (=send a single "\n").

On the Hipox Device it seems to take a lot longer for the password check to
return. Therefore a second prompt request is sent, when the first does not
return anything.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>