dev.py: Retreive IP addresses
A device now has a method to retreive its IP addresses. The result is basically
what felt best after some time of experimentation.

At first it should have just returned one single address, but the problem with
that approach is, that it really can have two different networks, at least at
DFE GmbH, and there is no way for a program to know which network should be
used in the test. Therefore everything is returned and it is up to the user
what to do about it.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>