0.15: dict references, extendable parsers, general testlogger handling
This is a little bigger release than planned first.

It all started with the requirement of a user to have conns and devs in
dictionaries, to have more features calling them, like setting default values
if none is found.

To apply that through the whole package line down to our tests other changes
were needed as well. One is that I wanted for a long time to have a more
flexible parser handling. Beside the default parsers for connections and
devices here we also have some classes that are specific to our internal
tooling. These need to be parsed and defined in a closed repository. With the
old parsing method it was not possible to have all of that outside of MONK. Now
we have done that and can go faster at providing indivual configuration options
in fixture files.

The last big point that changed is that the testlogger now is available
everywhere. Every MONK object has a testlogger attribute and a testlog method.
That way a testcase can separate its test specific log messages from the ones
that MONK itself produces and then it can be filtered.

Then there are also a lot of minor changes. One example would be that all the
things that belong to all MONK objects are in a single class now and most
classes here now extend that base class.

Signed-off-by: Erik Bernoth <erik.bernoth@gmail.com>